chore(core): restructure messaging service
