+++ src/bp/core/api.ts
@@ -7,6 +7,7 @@ import { SessionIdFactory } from 'core/dialog/sessions'
 import { EventEngine, EventRepository } from 'core/events'
 import { KeyValueStore } from 'core/kvs'
 import { LoggerProvider } from 'core/logger'
+import { ConversationService, MessageService } from 'core/messaging'
 import { ModuleLoader } from 'core/modules'
 import { WellKnownFlags } from 'core/sdk/enums'
 import { TYPES } from 'core/types'
@@ -21,7 +22,6 @@ import { isTrainingAlreadyStarted, isTrainingCanceled } from 'nlu-core/errors'
 import modelIdService from 'nlu-core/model-id-service'
 
 import { container } from './app.inversify'
-
 import { getMessageSignature } from './misc/security'
 import { Event, RealTimePayload } from './sdk/impl'
 import HTTPServer from './server'
@@ -31,8 +31,6 @@ import { DialogEngine } from './services/dialog/dialog-engine'
 import { HookService } from './services/hook/hook-service'
 import { JobService } from './services/job-service'
 import { MediaServiceProvider } from './services/media'
-import { ConversationService } from './services/messaging/conversations'
-import { MessageService } from './services/messaging/messages'
 import { NotificationsService } from './services/notification/service'
 import RealtimeService from './services/realtime'
 import { WorkspaceService } from './services/workspace-service'
