+++ src/bp/core/converse/converse-service.ts
@@ -1,6 +1,7 @@
 import { IO } from 'botpress/sdk'
 import { ConfigProvider } from 'core/config/config-loader'
 import { EventEngine } from 'core/events'
+import { MessageService, ConversationService } from 'core/messaging'
 import { TYPES } from 'core/types'
 import { ChannelUserRepository } from 'core/users'
 import { InvalidParameterError } from 'errors'
@@ -9,9 +10,6 @@ import { inject, injectable, postConstruct } from 'inversify'
 import { AppLifecycle, AppLifecycleEvents } from 'lifecycle'
 import _ from 'lodash'
 import ms from 'ms'
-
-import { ConversationService } from '../services/messaging/conversations'
-import { MessageService } from '../services/messaging/messages'
 
 export const converseApiEvents = new EventEmitter2()
