+++ src/bp/core/database/tables/bot-specific/index.ts
@@ -1,5 +1,3 @@
 export * from './ghost_content'
 export * from './notifications'
-export * from './conversations'
-export * from './messages'
 export * from './tasks'
