+++ src/bp/core/database/tables/index.ts
@@ -2,6 +2,8 @@ import { DialogSessionTable } from 'core/dialog/sessions/dialog_sessions-table'
 import { EventsTable } from 'core/events/event-table'
 import { KeyValueStoreTable } from 'core/kvs/kvs-table'
 import { LogsTable } from 'core/logger/logs-table'
+import { ConversationsTable } from 'core/messaging/conversations-table'
+import { MessagesTable } from 'core/messaging/messages-table'
 import { TelemetryTable } from 'core/telemetry/telemetry-table'
 import {
   WorkspaceInviteCodesTable,
@@ -14,14 +16,7 @@ import Knex from 'knex'
 
 import { Table } from '../interfaces'
 
-import {
-  ConversationsTable,
-  MessagesTable,
-  GhostFilesTable,
-  GhostRevisionsTable,
-  NotificationsTable,
-  TasksTable
-} from './bot-specific'
+import { GhostFilesTable, GhostRevisionsTable, NotificationsTable, TasksTable } from './bot-specific'
 import { MigrationsTable, ServerMetadataTable } from './server-wide'
 
 const tables: typeof Table[] = [
