+++ src/bp/core/messaging/index.ts
@@ -1,0 +1,4 @@
+export * from './conversations-repository'
+export * from './conversations-service'
+export * from './messages-repository'
+export * from './messages-service'
