+++ src/bp/core/services/services.inversify.ts
@@ -5,6 +5,7 @@ import { ConverseService } from 'core/converse'
 import { EventEngine, Queue, MemoryQueue } from 'core/events'
 import { KeyValueStore } from 'core/kvs'
 import { LogsJanitor, LogsService } from 'core/logger'
+import { MessageService, ConversationService } from 'core/messaging'
 import { DialogContainerModule } from 'core/services/dialog/dialog.inversify'
 import { CEJobService, JobService } from 'core/services/job-service'
 import { StatsService } from 'core/telemetry'
@@ -24,8 +25,6 @@ import { HintsService } from './hints'
 import { HookService } from './hook/hook-service'
 import CELicensingService from './licensing'
 import { MediaServiceProvider } from './media'
-import { ConversationService } from './messaging/conversations'
-import { MessageService } from './messaging/messages'
 import { CEMonitoringService, MonitoringService } from './monitoring'
 import { NLUService } from './nlu/nlu-service'
 import { NotificationsService } from './notification/service'
